app/backend/pyproject.toml
[project]
name = "car-ai-demo-backend"
version = "0.1.0"
description = "Car AI Demo Backend - FastAPI application for vehicle recommendations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "databricks-sql-connector>=3.0.0",
    "databricks-sdk>=0.30.0",
    "openai>=1.52.0",
    "mlflow>=2.15.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
